python/requirements.txt
numpy>=1.18.0
pymongo>=4.0.0
flask>=2.0.0
open3d>=0.15.0
matplotlib>=3.4.0
scipy>=1.7.0
pillow>=8.0.0
pathlib>=1.0.1
werkzeug>=2.0.0